# pizza_1337_preact/api_scanner_typescript/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/pizza-types": {
            "get": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Get All Pizza Types",
                "operationId": "get_all_pizza_types_v1_pizza_types_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get All Pizza Types V1 Pizza Types Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PizzaTypeSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Create Pizza Type",
                "operationId": "create_pizza_type_v1_pizza_types_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PizzaTypeCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaTypeSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/pizza-types/{pizza_type_id}": {
            "get": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Get Pizza Type",
                "operationId": "get_pizza_type_v1_pizza_types__pizza_type_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Pizza Type Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "pizza_type_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaTypeSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Update Pizza Type",
                "operationId": "update_pizza_type_v1_pizza_types__pizza_type_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Pizza Type Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "pizza_type_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PizzaTypeCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaTypeSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Delete Pizza Type",
                "operationId": "delete_pizza_type_v1_pizza_types__pizza_type_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Pizza Type Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "pizza_type_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/pizza-types/{pizza_type_id}/dough": {
            "get": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Get Pizza Type Dough",
                "operationId": "get_pizza_type_dough_v1_pizza_types__pizza_type_id__dough_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Pizza Type Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "pizza_type_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoughSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/order/{order_id}/pizzas": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Get Pizzas From Order",
                "operationId": "get_pizzas_from_order_v1_order__order_id__pizzas_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Order Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "order_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get Pizzas From Order V1 Order  Order Id  Pizzas Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JoinedPizzaPizzaTypeSchema"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Add Pizza To Order",
                "operationId": "add_pizza_to_order_v1_order__order_id__pizzas_post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Order Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "order_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PizzaCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaWithoutPizzaTypeSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "order"
                ],
                "summary": "Delete Pizza From Order",
                "operationId": "delete_pizza_from_order_v1_order__order_id__pizzas_delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Order Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "order_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PizzaWithoutPizzaTypeSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DoughSchema": {
                "title": "DoughSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "stock",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "stock": {
                        "title": "Stock",
                        "type": "integer"
                    },
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "JoinedPizzaPizzaTypeSchema": {
                "title": "JoinedPizzaPizzaTypeSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
...
            "PizzaCreateSchema": {
                "title": "PizzaCreateSchema",
                "required": [
                    "pizza_type_id"
                ],
                "type": "object",
                "properties": {
                    "pizza_type_id": {
                        "title": "Pizza Type Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "PizzaTypeCreateSchema": {
                "title": "PizzaTypeCreateSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "dough_id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "dough_id": {
                        "title": "Dough Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "PizzaTypeSchema": {
                "title": "PizzaTypeSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
...
            "PizzaWithoutPizzaTypeSchema": {
                "title": "PizzaWithoutPizzaTypeSchema",
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    },
...
}